ai_vc_debate/requirements.txt
gradio
openai-agents
httpx
python-dotenv
pydantic
